scripts/gta_sfm/requirements.txt
Pillow
numpy
h5py
